feat(reporter): move types into reporter file
